Fix NumberFormatException in Learning Paths
Changed progress_percentage from Int to Double in LearningPath model
to match backend decimal:2 type, fixing NumberFormatException when
parsing API responses.

Changes:
- Task.kt: Changed progress_percentage type from Int to Double
- PathsViewModel.kt: Added parseDoubleSafe() helper method
- PathsViewModel.kt: Use parseDoubleSafe() for progress_percentage parsing
- PathsViewModel.kt: Convert Double to Int for overall progress calculation
- PathsAdapter.kt: Convert progress_percentage to Int for display
- LearningPathDetailActivity.kt: Convert progress_percentage to Int for UI

+++ mobileandroid/app/src/main/java/ecccomp/s2240788/mobile_android/data/models/Task.kt
@@ -124,7 +124,7 @@ data class LearningPath(
     val description: String?,
     val category: String?,
     val status: String, // active, completed, paused
-    val progress_percentage: Int,
+    val progress_percentage: Double, // Changed from Int to Double to match backend decimal:2 type
     val total_milestones: Int,
     val completed_milestones: Int,
     val target_date: String?,
